Refactor: Extension 분리
#28
