+++ Bookjam/Extension/UIView+.swift
@@ -1,0 +1,17 @@
+//
+//  UIView+.swift
+//  Bookjam
+//
+//  Created by YOUJIM on 2023/08/07.
+//
+
+import UIKit
+
+extension UIView {
+   func roundCorners(corners: UIRectCorner, radius: CGFloat) {
+        let path = UIBezierPath(roundedRect: bounds, byRoundingCorners: corners, cornerRadii: CGSize(width: radius, height: radius))
+        let mask = CAShapeLayer()
+        mask.path = path.cgPath
+        layer.mask = mask
+    }
+}
