+++ Bookjam/Extension/UIViewController+.swift
@@ -1,0 +1,93 @@
+//
+//  UIViewController+.swift
+//  Bookjam
+//
+//  Created by YOUJIM on 2023/08/07.
+//
+
+import SwiftUI
+import UIKit
+
+
+// MARK: - 화면 여백 공간 눌렀을 때 키보드 숨길 수 있게 하는 extension
+// hideKeyboard() 호출해서 사용함
+
+extension UIViewController {
+    func hideKeyboard() {
+        let tap: UITapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(UIViewController.dismissKeyboard))
+        view.addGestureRecognizer(tap)
+    }
+    
+    @objc func dismissKeyboard() {
+        view.endEditing(true)
+    }
+    
+    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+        textField.resignFirstResponder()
+        return true
+    }
+}
+
+
+// MARK: 프리뷰 extenstion
+
+#if canImport(SwiftUI) && DEBUG
+    import SwiftUI
+
+    public struct UIViewPreview<View: UIView>: UIViewRepresentable {
+        public let view: View
+        public init(_ builder: @escaping () -> View) {
+            view = builder()
+        }
+        // MARK: - UIViewRepresentable
+        public func makeUIView(context: Context) -> UIView {
+            return view
+        }
+        public func updateUIView(_ view: UIView, context: Context) {
+            view.setContentHuggingPriority(.defaultHigh, for: .horizontal)
+            view.setContentHuggingPriority(.defaultHigh, for: .vertical)
+        }
+    }
+
+    public struct UIViewControllerPreview<ViewController: UIViewController>: UIViewControllerRepresentable {
+        public let viewController: ViewController
+
+        public init(_ builder: @escaping () -> ViewController) {
+            viewController = builder()
+        }
+
+        // MARK: - UIViewControllerRepresentable
+        public func makeUIViewController(context: Context) -> ViewController {
+            viewController
+        }
+
+        @available(iOS 13.0, tvOS 13.0, *)
+        @available(OSX, unavailable)
+        @available(watchOS, unavailable)
+        public func updateUIViewController(_ uiViewController: ViewController, context: UIViewControllerRepresentableContext<UIViewControllerPreview<ViewController>>) {
+            return
+        }
+    }
+#endif
+
+
+// MARK: ViewController 말고 View나 Cell도 프리뷰 가능하게 해주는 Extension
+
+#if DEBUG
+extension UIViewController {
+    private struct Preview: UIViewControllerRepresentable {
+            let viewController: UIViewController
+
+            func makeUIViewController(context: Context) -> UIViewController {
+                return viewController
+            }
+
+            func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
+            }
+        }
+
+        func toPreview() -> some View {
+            Preview(viewController: self)
+        }
+}
+#endif
